refactored some tests
